Reorganizes Navbar into own directory w/ subcomponents
